CIS-3016 README Update
Updated the working in README.md to rein in use of the word
"description" unless it is specifically referring to the "Description"
tab or column names in measures.csv. Preferred synonyms were "summary",
"definition", and "overview".

While updating the file, also updated column definitions indicating
Markdown content to reflect the decision to use HTML.